Removed C++ Code - OOP is dead. Long live prodedural programming.
